Day 3 Project - Treasure Island
The game is a basic version of classic treasure island game.
Here, the player is asked options to select at each step of the game to proceed further. If all the options are correct, the player has won the game and the treasure.
Methods used: Conditional statements, Nested If else,  multiple if statements. Basic graphics drawing by using ASCII values.